Support URLs and file paths for reading configs (#6)
Rather than specifying the project name and using a flag, we'll cascade a set of options.  First, we'll attempt to read a file.  If the file doesn't exist, then we'll attempt to open a url.  If the url isn't opened, we'll create a url to the build configuration directory and assume that the first parameter is the project name.